move auth to separate file, add integration test for redirect
